minor fixes and refactoring (added private.h)
