+++ src/plugins/memdump/private.h
@@ -102,127 +102,94 @@
  *                                                                         *
  ***************************************************************************/
 
-#include <config.h>
-#include <libvmi/libvmi.h>
+#ifndef MEMDUMP_PRIVATE_H
+#define MEMDUMP_PRIVATE_H
 
-#include "memdump.h"
-#include "private.h"
+template<typename T>
+struct copy_on_write_result_t: public call_result_t<T>
+{
+    copy_on_write_result_t(T* src) : call_result_t<T>(src), vaddr(), pte(), old_cow_pa() {}
+
+    addr_t vaddr;
+    addr_t pte;
+    addr_t old_cow_pa;
+};
+
+template<typename T>
+struct map_view_of_section_result_t: public call_result_t<T>
+{
+    map_view_of_section_result_t(T* src) : call_result_t<T>(src), section_handle(), process_handle(), base_address_ptr() {}
+
+    uint64_t section_handle;
+    uint64_t process_handle;
+    addr_t base_address_ptr;
+};
+
+enum target_hook_state
+{
+    HOOK_FIRST_TRY,
+    HOOK_PAGEFAULT_RETRY,
+    HOOK_FAILED,
+    HOOK_OK
+};
+
+typedef event_response_t (*callback_t)(drakvuf_t drakvuf, drakvuf_trap_info* info);
+class memdump;
+
+struct hook_target_entry_t
+{
+    vmi_pid_t pid;
+    std::string target_name;
+    callback_t callback;
+    target_hook_state state;
+    drakvuf_trap_t* trap;
+    memdump* plugin;
+
+    hook_target_entry_t(std::string target_name, callback_t callback, memdump* plugin)
+            : target_name(target_name), callback(callback), state(HOOK_FIRST_TRY), plugin(plugin) {}
+};
+
+struct user_dll_t
+{
+    // relevant while loading
+    addr_t dtb;
+    uint32_t thread_id;
+    addr_t real_dll_base;
+    bool is_hooked;
+
+    // internal, for page faults
+    addr_t pf_current_addr;
+    addr_t pf_max_addr;
+
+    // one entry per hooked function
+    std::vector<hook_target_entry_t> targets;
+};
+
+struct target_config_entry_t
+{
+    std::string dll_name;
+    std::string function_name;
+};
+
+// type of a pointer residing on stack
+enum sptr_type_t
+{
+    ERROR,   // problem with stack inspection
+    MAIN,    // pointer to a main module
+    LINKED,  // pointer to some linked DLL module
+    UNLINKED // pointer to some non-legit memory
+};
 
 sptr_type_t check_module_linked_wow(drakvuf_t drakvuf,
                                     vmi_instance_t vmi,
                                     memdump* plugin,
                                     drakvuf_trap_info_t* info,
-                                    addr_t dll_base)
-{
-    access_context_t ctx =
-    {
-        .translate_mechanism = VMI_TM_PROCESS_DTB,
-        .dtb = info->regs->cr3
-    };
-
-    addr_t wow_peb = drakvuf_get_wow_peb(drakvuf, &ctx, info->proc_data.base_addr);
-
-    if (!wow_peb)
-        return ERROR;
-
-    addr_t module_list_head;
-
-    if (!drakvuf_get_module_list_wow(drakvuf, &ctx, wow_peb, &module_list_head))
-        return ERROR;
-
-    addr_t next_module = module_list_head;
-    bool is_first = true;
-    sptr_type_t ret = UNLINKED;
-
-    while (1)
-    {
-        uint32_t tmp_next = 0;
-        ctx.addr = next_module;
-        if (VMI_FAILURE == vmi_read_32(vmi, &ctx, &tmp_next))
-        {
-            ret = ERROR;
-            break;
-        }
-
-        if (module_list_head == (addr_t)tmp_next || !tmp_next)
-            break;
-
-        uint32_t tmp_dll_base;
-        ctx.addr = next_module + plugin->dll_base_wow_rva;
-        if (vmi_read_32(vmi, &ctx, &tmp_dll_base) == VMI_SUCCESS)
-        {
-            if (dll_base == (addr_t)tmp_dll_base)
-            {
-                ret = LINKED;
-                break;
-            }
-        }
-
-        next_module = (addr_t)tmp_next;
-        is_first = false;
-    }
-
-    if (is_first && ret == LINKED)
-        ret = MAIN;
-
-    return ret;
-}
+                                    addr_t dll_base);
 
 sptr_type_t check_module_linked(drakvuf_t drakvuf,
                                 vmi_instance_t vmi,
                                 memdump* plugin,
                                 drakvuf_trap_info_t* info,
-                                addr_t dll_base)
-{
-    sptr_type_t sub_ret = check_module_linked_wow(drakvuf, vmi, plugin, info, dll_base);
+                                addr_t dll_base);
 
-    if (sub_ret != ERROR && sub_ret != UNLINKED)
-        return sub_ret;
-
-    addr_t module_list_head;
-    if (!drakvuf_get_module_list(drakvuf, info->proc_data.base_addr, &module_list_head))
-        return ERROR;
-
-    addr_t next_module = module_list_head;
-    access_context_t ctx =
-    {
-        .translate_mechanism = VMI_TM_PROCESS_DTB,
-        .dtb = info->regs->cr3
-    };
-
-    bool is_first = true;
-    sptr_type_t ret = UNLINKED;
-
-    while (1)
-    {
-        addr_t tmp_next = 0;
-        ctx.addr = next_module;
-        if (VMI_FAILURE == vmi_read_addr(vmi, &ctx, &tmp_next))
-        {
-            ret = ERROR;
-            break;
-        }
-
-        if (module_list_head == tmp_next || !tmp_next)
-            break;
-
-        addr_t tmp_dll_base;
-        ctx.addr = next_module + plugin->dll_base_rva;
-        if (vmi_read_addr(vmi, &ctx, &tmp_dll_base) == VMI_SUCCESS)
-        {
-            if (dll_base == tmp_dll_base)
-            {
-                ret = LINKED;
-                break;
-            }
-        }
-
-        is_first = false;
-        next_module = tmp_next;
-    }
-
-    if (is_first && ret == LINKED)
-        ret = MAIN;
-
-    return ret;
-}
+#endif
