+++ src/plugins/memdump/userhook.cpp
@@ -123,11 +123,13 @@
 #include <libvmi/libvmi.h>
 #include <libvmi/peparse.h>
 #include <assert.h>
+
 #include "memdump.h"
+#include "private.h"
 
 static event_response_t usermode_hook_cb(drakvuf_t drakvuf, drakvuf_trap_info* info)
 {
-    hook_target_entry* target = (hook_target_entry*)info->trap->data;
+    hook_target_entry_t* target = (hook_target_entry_t*)info->trap->data;
 
     if (target->pid != info->proc_data.pid)
         return VMI_EVENT_RESPONSE_NONE;
@@ -337,7 +339,7 @@ static user_dll_t* create_dll_meta(drakvuf_t drakvuf, drakvuf_trap_info* info, m
     return &it->second.back();
 }
 
-static bool make_trap(drakvuf_t drakvuf, drakvuf_trap_info* info, hook_target_entry* target, addr_t exec_func)
+static bool make_trap(drakvuf_t drakvuf, drakvuf_trap_info* info, hook_target_entry_t* target, addr_t exec_func)
 {
     target->pid = info->proc_data.pid;
